+++ src/controllers/Lane.js
@@ -170,7 +170,7 @@ class Lane extends Component {
           key={card.id}
           index={idx}
           style={card.style || cardStyle}
-          className="react-trello-card"
+          className="kanban-board-card"
           onDelete={onDeleteCard}
           onClick={e => this.handleCardClick(e, card)}
           onChange={updatedCard => this.updateCard(updatedCard)}
@@ -258,7 +258,7 @@ class Lane extends Component {
       onCardMoveAcrossLanes,
       ...otherProps
     } = this.props
-    const allClassNames = classNames('react-trello-lane', this.props.className || '')
+    const allClassNames = classNames('kanban-board-lane', this.props.className || '')
     const showFooter = collapsibleLanes && cards.length > 0
     return (
       <components.Section